\name{vsore_fit}
\alias{vsore_fit}
\title{Alias for vsore (backward compatibility)}
\usage{vsore_fit(z, R, anchor_weights, n, ...)}
\arguments{
\item{z, R, anchor_weights, n, ...}{See \code{\link{vsore}}.}
}
\value{See \code{\link{vsore}}.}
\description{
Backward-compatible alias retained for scripts written against earlier
versions of the package.
}
\seealso{\code{\link{vsore}}}
